Intermediate commit: will not compile successfully.
Changed runtax.Rule completely:
-Removed all instance chained building methods.
-Added enum Type describing the type of the Rule (ie.: Sequence, Regex, Text, OneOf, Optional, OneOrMore, ZeroOrMore, and so on).
-Regex is now assembled using the type instead of special characters added to inner Rules in the respective methods.
-Renamed runtax.Rule.RuleEntry to runtax.Rule.Entry.
-Removed #name(String).
-Added runtax.Rule.Entry based methods to runtax.Rule.Builder, which are now the main building methods, while the runtax.Rule based ones are now dependent on those.
-Old version still exists under runtax.OldRule for now.

Added runtax.ParsedRuleMap for parsing a runtax.RuleMap from the contents of a File / a String.

Added runtax.ParserRuleMap describing a RuleMap for the default syntax for the ParsedRuleMap.

Added runtax.Parser for parsing the contents of a File / a String into an AST (technically not Abstract, just a Syntax Tree).